python_service/requirements.txt
flask
flask-cors
transformers
torch
nltk
google-cloud-vision
